+++ framework/global/modularity/modularity.cmake
@@ -4,6 +4,4 @@ set (MODULARITY_SRC
     ${CMAKE_CURRENT_LIST_DIR}/modulesioc.h
     ${CMAKE_CURRENT_LIST_DIR}/imoduleexport.h
     ${CMAKE_CURRENT_LIST_DIR}/ioc.h
-    ${CMAKE_CURRENT_LIST_DIR}/serviceinjector.h
-    ${CMAKE_CURRENT_LIST_DIR}/servicesresolver.h
 )
