+++ framework/telemetry/CMakeLists.txt
@@ -16,38 +16,33 @@
 #  along with this program; if not, write to the Free Software
 #  Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
 #=============================================================================
-get_filename_component(DIR_NAME ${CMAKE_CURRENT_SOURCE_DIR} NAME)
-message("Configuring " ${DIR_NAME})
 
-find_package(Qt5 COMPONENTS Core Gui REQUIRED CONFIG)
-add_definitions(
-      ${Qt5Core_DEFINITIONS}
-      ${Qt5Gui_DEFINITIONS}
-      )
+set(MODULE telemetry)
+set(LIBRARY_TYPE STATIC)
+message(STATUS "Configuring " ${MODULE})
+
 include_directories(
-      ${Qt5Core_INCLUDE_DIRS}
-      ${Qt5Gui_INCLUDE_DIRS}
       ${PROJECT_SOURCE_DIR}/thirdparty
       ${PROJECT_SOURCE_DIR}/mscore
       ${PROJECT_SOURCE_DIR}/framework/global
       ${PROJECT_BINARY_DIR}
       )
 
-set(CMAKE_AUTOMOC ON)
-set(CMAKE_INCLUDE_CURRENT_DIR ON)
-
-set(LIBRARY_TYPE STATIC)
-
 file(GLOB_RECURSE telemetry_SOURCES RELATIVE ${CMAKE_CURRENT_LIST_DIR} *.cpp)
 file(GLOB_RECURSE telemetry_HEADERS RELATIVE ${CMAKE_CURRENT_LIST_DIR} *.h)
 
 qt5_add_resources(RCC_SOURCES telemetry_resources.qrc)
 
-add_library(telemetry ${LIBRARY_TYPE} ${telemetry_SOURCES} ${telemetry_HEADERS} ${RCC_SOURCES})
-target_link_libraries(telemetry global google_analytics)
+add_library(${MODULE} ${LIBRARY_TYPE}
+    ${telemetry_SOURCES}
+    ${telemetry_HEADERS}
+    ${RCC_SOURCES}
+)
+
+target_link_libraries(${MODULE} global google_analytics)
 
 if (MSVC)
-  set_target_properties( telemetry PROPERTIES
+  set_target_properties (${MODULE} PROPERTIES
     COMPILE_FLAGS "/wd4127"
   )
 endif (MSVC)
