+++ framework/telemetry/qml/DialogButton.qml
