+++ framework/telemetry/qml/TelemetryPermissionDialog.qml
