+++ framework/telemetry/qml/TextLabel.qml
