+++ framework/telemetry/telemetry_resources.qrc
