+++ framework/telemetry/widgets/telemetrypermissiondialog.cpp
