+++ framework/telemetry/widgets/telemetrypermissiondialog.h
